Extracted sub components and moved utils to separate file
